Removed Repositories all together with the prototype Brands data and controller.
