+++ Source/InvestorAPI/InvestorAPI.Core/Enums/DeleteResult.cs
@@ -1,0 +1,9 @@
+﻿namespace InvestorAPI.Core
+{
+    public enum DeleteResult
+    {
+        Success = 0,
+        Failed = 1,
+        EntityNotFound = 2
+    }
+}
