+++ Source/InvestorAPI/InvestorAPI.Core/Enums/OperationError.cs
@@ -1,0 +1,12 @@
+﻿namespace InvestorAPI.Core
+{
+    public enum OperationError
+    {
+        None = 0,
+        EntityNotFound = 1,
+        UnprocessableEntity = 2,
+        ValidationError = 3,
+        DatabaseError = 4,
+        ExternalError = 5
+    }
+}
