--- Source/InvestorData/InvestorData/Repositories/Abstraction/Base/Enums.cs
@@ -1,16 +1,0 @@
-﻿namespace InvestorData
-{
-    public enum LoginResult
-    {
-        Success,
-        Failed,
-        UserNotFound
-    }
-    
-    public enum DeleteResult
-    {
-        Success,
-        Failed,
-        EntityNotFound
-    }
-}
