--- Source/InvestorData/InvestorData/Repositories/Abstraction/Management/IBusinessRepository.cs
@@ -1,13 +1,0 @@
-﻿namespace InvestorData
-{
-
-    /// <summary>
-    /// Provides an abstraction for a repository which manages <see cref="Business"/> entities.
-    /// </summary>
-    public interface IBusinessRepository : IRepository<Business>
-    {
-
-
-
-    }
-}
