--- Source/InvestorData/InvestorData/Repositories/Abstraction/Management/IBusinessTypeRepository.cs
@@ -1,13 +1,0 @@
-﻿namespace InvestorData
-{
-
-    /// <summary>
-    /// Provides an abstraction for a repository which manages <see cref="BusinessType"/> entities.
-    /// </summary>
-    public interface IBusinessTypeRepository : IRepository<BusinessType>
-    {
-
-
-
-    }
-}
