--- Source/InvestorData/InvestorData/Repositories/Management/BusinessTypeRepository.cs
@@ -1,36 +1,0 @@
-﻿namespace InvestorData
-{
-
-    /// <summary>
-    /// A repository that manages <see cref="BusinessType"/> entities.
-    /// </summary>
-    internal class BusinessTypeRepository<TContext> : Repository<BusinessType>, IBusinessTypeRepository where TContext : InvestorDbContext
-    {
-
-        #region Constructor
-
-        public BusinessTypeRepository(TContext dbContext) : base(dbContext, dbContext.BusinessTypes)
-        {
-
-        }
-
-        #endregion
-
-
-        #region Abstract Implementation
-
-        protected override IQueryable<BusinessType> QueryIncludingMinimalData()
-        {
-            return DbSet;
-        }
-
-        protected override IQueryable<BusinessType> QueryIncludingFullData()
-        {
-            return QueryIncludingMinimalData();
-        }
-
-        #endregion
-
-    }
-
-}
